djangoBlog/requirements.txt
Django==2.0.5
django-contrib-comments==1.8.0
django-crispy-forms==1.7.2
django-ipware==2.1.0
django-mdeditor==0.1.4
django-taggit==0.22.2
django-threadedcomments==1.2
django-uuslug==1.1.8
docutils==0.14
Markdown==2.6.11
Pillow==5.1.0
python-slugify==1.2.5
pytz==2018.4
six==1.11.0
Unidecode==1.0.22
